.github/scripts/requirements.txt
cloudinary>=1.40.0
python-frontmatter>=1.1.0
playwright>=1.44.0
requests>=2.32.0
